Got search page working, though api db query not working
