+++ client/components/Searchbar.js
@@ -2,6 +2,7 @@ import React from 'react'
 
 const Searchbar = props => {
   const {handleSubmit, query, handleChange} = props
+  console.log("HELLOOO")
   return (
     <div>
       <form onSubmit={(event) => handleSubmit(event)}>
